# docs/swagger.yaml
basePath: /
definitions:
  requests.CalculateTotalCost:
    properties:
      end_period:
        type: string
      service_name:
        type: string
      start_period:
        type: string
      user_id:
        type: string
    required:
    - end_period
    - start_period
    type: object
  requests.SubRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  responses.CalculateTotalCost:
    properties:
      total:
        type: integer
    required:
    - total
    type: object
  responses.SubResponse:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - id
    - price
    - service_name
    - start_date
    - user_id
    type: object
host: localhost:8080
info:
  contact: {}
  description: service for subscriptions users
  title: Subscription Service
  version: 0.0.1
paths:
  /subscriptions:
    get:
      description: Возвращает список подписок с поддержкой пагинации
      parameters:
      - default: 10
        description: Количество подписок на странице
        in: query
        name: limit
        type: integer
      - default: 0
        description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.SubResponse'
            type: array
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: Получение списка подписок
    post:
      consumes:
      - application/json
      description: Создание новой записи подписки
      parameters:
      - description: структура запроса
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/requests.SubRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.SubResponse'
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: Создание подписки
      tags:
      - subscriptions
  /subscriptions/{sub_id}:
    delete:
      description: Удаление подписки по ID
      parameters:
      - description: path format
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "404":
          description: подписка не найдена
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: Удаление подписки
      tags:
      - subscriptions
    get:
      description: Запрос на получение подписки по ее ID
      parameters:
      - description: path format
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SubResponse'
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "404":
          description: подписка не найдена
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: Запрос на получение подписки
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновление подписки по ID
      parameters:
      - description: path format
        in: path
        name: id
        required: true
        type: string
      - description: структура запроса
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/requests.SubRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SubResponse'
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "404":
          description: подписка не найдена
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: Обновление подписки
      tags:
      - subscriptions
  /subscriptions/total:
    post:
      consumes:
      - application/json
      description: Расчет общей стоимость подписок за определенный период с использованием
        дополнительных фильтров
      parameters:
      - description: структура запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CalculateTotalCost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.CalculateTotalCost'
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: Рассчет общую стоимость подписки
      tags:
      - subscriptions
swagger: "2.0"